apps/dairyflow/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sse-starlette==2.1.3
databricks-sdk==0.40.0
pydantic==2.10.3
aiofiles==24.1.0
